Split up model.py
Splits it up into model.user, model.room, model.file, model.exc
