+++ sogs/model/room.py
@@ -1,116 +1,16 @@
-from __future__ import annotations
-
-from typing import Optional, Union
-import flask
-import sqlalchemy.exc
-
-from . import config
-from . import db
-from . import utils
-from . import crypto
-from . import http
-from .omq import send_mule
-from .web import app
-from .db import query
+from .. import config, crypto, db, utils
+from ..db import query
+from ..omq import send_mule
+from ..web import app
+from .user import User
+from .file import File
+from .exc import NoSuchRoom, NoSuchFile, AlreadyExists, BadPermission, PostRejected, PostRateLimited
 
 import os
 import re
+import sqlalchemy.exc
 import time
-
-
-if config.PROFANITY_FILTER:
-    import better_profanity
-
-    if config.PROFANITY_CUSTOM:
-        better_profanity.profanity.load_censor_words_from_file(config.PROFANITY_CUSTOM)
-    else:
-        better_profanity.profanity.load_censor_words()
-
-
-class NotFound(LookupError):
-    """Base class for NoSuchRoom, NoSuchFile, etc."""
-
-    pass
-
-
-class NoSuchRoom(NotFound):
-    """Thrown when trying to construct a Room from a token that doesn't exist"""
-
-    def __init__(self, token):
-        self.token = token
-        super().__init__(f"No such room: {token}")
-
-
-class NoSuchFile(NotFound):
-    """Thrown when trying to construct a File from a token that doesn't exist"""
-
-    def __init__(self, id):
-        self.id = id
-        super().__init__(f"No such file: {id}")
-
-
-class NoSuchUser(NotFound):
-    """Thrown when attempting to retrieve a user that doesn't exist and auto-vivification of the
-    user room is disabled"""
-
-    def __init__(self, session_id):
-        self.session_id = session_id
-        super().__init__(f"No such user: {session_id}")
-
-
-class AlreadyExists(RuntimeError):
-    """
-    Thrown when attempting to create a record (e.g. a Room) that already exists.
-
-    e.type is the type object (e.g. sogs.model.Room) that could not be constructed, if applicable.
-    e.value is the unique value that already exists (e.g. the room token), if applicable.
-    """
-
-    def __init__(self, msg, type=None, value=None):
-        super().__init__(msg)
-        self.type = type
-        self.value = value
-
-
-class BadPermission(RuntimeError):
-    """Thrown when attempt to perform an action that the given user does not have permission to do;
-    for example, attempting to delete someone else's posts when not a moderator."""
-
-    def __init__(self, msg=None):
-        super().__init__("Permission denied" if msg is None else msg)
-
-
-class PostRejected(RuntimeError):
-    """
-    Thrown when a post is refused for some reason other than a permission error (e.g. the post
-    contains bad words)
-    """
-
-    def __init__(self, msg=None):
-        super().__init__("Post rejected" if msg is None else msg)
-
-
-class PostRateLimited(PostRejected):
-    """Thrown when attempting to post too frequently in a room"""
-
-    def __init__(self, msg=None):
-        super().__init__("Rate limited" if msg is None else msg)
-
-
-# Map uncaught model exceptions into flask http exceptions
-@app.errorhandler(NotFound)
-def abort_bad_room(e):
-    flask.abort(http.NOT_FOUND)
-
-
-@app.errorhandler(BadPermission)
-def abort_perm_denied(e):
-    flask.abort(http.FORBIDDEN)
-
-
-@app.errorhandler(PostRejected)
-def abort_post_rejected(e):
-    flask.abort(http.TOO_MANY_REQUESTS)
+from typing import Optional, Union
 
 
 class Room:
@@ -1197,281 +1097,6 @@ class Room:
             raise
 
 
-class File:
-    """
-    Class representing a user stored in the database.
-
-    Properties:
-        id - the numeric file id, i.e. primary key
-        room - the Room that this file belongs to (only retrieved on demand).
-        uploader - the User that uploaded this file (only retrieved on demand).
-        size - the size (in bytes) of this file
-        uploaded - unix timestamp when the file was uploaded
-        expiry - unix timestamp when the file expires.  None for non-expiring files.
-        path - the path of this file on disk, relative to the base data directory.
-        filename - the suggested filename provided by the user.  None for there is no suggestion
-            (this will always be the case for files uploaded by legacy Session clients).
-    """
-
-    def __init__(self, row=None, *, id=None):
-        """
-        Constructs a file from a pre-retrieved row *or* a file id.  Raises NoSuchFile if the id does
-        not exist in the database.
-        """
-        if sum(x is not None for x in (id, row)) != 1:
-            raise ValueError("File() error: exactly one of id/row is required")
-        if id is not None:
-            row = query("SELECT * FROM files WHERE id = :f", f=id).first()
-            if not row:
-                raise NoSuchFile(id)
-
-        (
-            self.id,
-            self._fetch_room_id,
-            self._fetch_uploader_id,
-            self.size,
-            self.uploaded,
-            self.expiry,
-            self.filename,
-            self.path,
-        ) = (
-            row[c]
-            for c in ('id', 'room', 'uploader', 'size', 'uploaded', 'expiry', 'filename', 'path')
-        )
-        self._uploader = None
-        self._room = None
-
-    @property
-    def room(self):
-        """
-        Accesses the Room in which this image is posted; this is fetched from the database the first
-        time this is accessed.  In theory this can return None if the Room is in the process of
-        being deleted but the Room's uploaded files haven't been deleted yet.
-        """
-        if self._fetch_room_id is not None:
-            try:
-                self._room = Room(id=self._fetch_room_id)
-            except NoSuchFile:
-                pass
-            self._fetch_room_id = None
-        return self._room
-
-    @property
-    def room_id(self):
-        """
-        Accesses the id of the room to which this file was uploaded.  Equivalent to .room.id, except
-        that we don't fetch/cache the Room row.
-        """
-        return self._fetch_room_id if self._room is None else self._fetch_room_id
-
-    @property
-    def uploader(self):
-        """
-        Accesses the User who uploaded this file.  Retrieves from the database the first time this
-        is accessed.
-        """
-
-        if self._fetch_uploader_id is not None:
-            try:
-                self._uploader = User(id=self._fetch_uploader_id)
-            except NoSuchUser:
-                pass
-            self._fetch_uploader_id = None
-        return self._uploader
-
-    @property
-    def uploader_id(self):
-        """
-        Accesses the id of the user who uploaded this file.  Equivalent to .uploader.id, except
-        that we don't fetch/cache the User row.
-        """
-        return self._fetch_uploader_id if self._uploader is None else self._uploader.id
-
-    def read(self):
-        """Reads the file from disk, as bytes."""
-        with open(self.path, 'rb') as f:
-            return f.read()
-
-    def read_base64(self):
-        """Reads the file from disk and encodes as base64."""
-        return utils.encode_base64(self.read())
-
-    def set_expiry(self, duration=None, forever=False):
-        """
-        Updates the file expiry to `duration` seconds from now, or to unlimited if `forever` is
-        True.  If duration is None (and not using forever) then the default expiry will be used.
-        """
-        expiry = (
-            None
-            if forever
-            else time.time()
-            + (duration if duration is not None else config.UPLOAD_DEFAULT_EXPIRY_DAYS)
-        )
-        query("UPDATE files SET expiry = :when WHERE id = :f", when=expiry, f=self.id)
-        self.expiry = expiry
-
-
-class User:
-    """
-    Class representing a user stored in the database.
-
-    Properties:
-        id - the database primary key for this user row
-        session_id - the session_id of the user, in hex
-        created - unix timestamp when the user was created
-        last_active - unix timestamp when the user was last active
-        banned - True if the user is (globally) banned
-        admin - True if the user is a global admin
-        moderator - True if the user is a global moderator
-        visible_mod - True if the user's admin/moderator status should be visible in rooms
-    """
-
-    def __init__(self, row=None, *, id=None, session_id=None, autovivify=True, touch=False):
-        """
-        Constructs a user from a pre-retrieved row *or* a session id or user primary key value.
-
-        autovivify - if True and we are given a session_id that doesn't exist, create a default user
-        row and use it to populate the object.  This is the default behaviour.  If False and the
-        session_id doesn't exist then a NoSuchUser is raised if the session id doesn't exist.
-
-        touch - if True (default is False) then update the last_activity time of this user before
-        returning it.
-        """
-
-        if sum(x is not None for x in (row, session_id, id)) != 1:
-            raise ValueError("User() error: exactly one of row/session_id/id is required")
-
-        self._touched = False
-        if session_id is not None:
-            row = query("SELECT * FROM users WHERE session_id = :s", s=session_id).first()
-
-            if not row and autovivify:
-                with db.transaction():
-                    query("INSERT INTO users (session_id) VALUES (:s)", s=session_id)
-                    row = query("SELECT * FROM users WHERE session_id = :s", s=session_id).first()
-                # No need to re-touch this user since we just created them:
-                self._touched = True
-
-        elif id is not None:
-            row = query("SELECT * FROM users WHERE id = :u", u=id).fetchone()
-
-        if row is None:
-            raise NoSuchUser(session_id if session_id is not None else id)
-
-        self.id, self.session_id, self.created, self.last_active = (
-            row[c] for c in ('id', 'session_id', 'created', 'last_active')
-        )
-        self.banned, self.global_moderator, self.global_admin, self.visible_mod = (
-            bool(row[c]) for c in ('banned', 'moderator', 'admin', 'visible_mod')
-        )
-
-        if touch:
-            with db.transaction():
-                self._touch()
-
-    def __str__(self):
-        """Returns string representation of a user: U[050123…cdef], the id prefixed with @ or % if
-        the user is a global admin or moderator, respectively."""
-        if len(self.session_id) != 66:
-            # Something weird (e.g. the "deleted" id from an old sogs import), just print directly
-            return f"U[{self.session_id}]"
-        return "U[{}{}…{}]".format(
-            '@' if self.global_admin else '%' if self.global_moderator else '',
-            self.session_id[:6],
-            self.session_id[-4:],
-        )
-
-    def _touch(self):
-        query(
-            """
-            UPDATE users SET last_active = :now
-            WHERE id = :u
-            """,
-            u=self.id,
-            now=time.time(),
-        )
-        self._touched = True
-
-    def touch(self, force=False):
-        """
-        Updates the last activity time of this user.  This method only updates the first time it is
-        called (and possibly not even then, if we auto-vivified the user row), unless `force` is set
-        to True.
-        """
-        if not self._touched or force:
-            with db.transaction():
-                self._touch()
-
-    def update_room_activity(self, room):
-        query(
-            """
-            INSERT INTO room_users ("user", room) VALUES (:u, :r)
-            ON CONFLICT("user", room) DO UPDATE
-            SET last_active = :now
-            """,
-            u=self.id,
-            r=room.id,
-            now=time.time(),
-        )
-
-    def set_moderator(self, *, added_by: User, admin=False, visible=False):
-        """
-        Make this user a global moderator or admin.  If the user is already a global mod/admin then
-        their status is updated according to the given arguments (that is, this can promote/demote).
-        """
-
-        if not added_by.global_admin:
-            app.logger.warning(
-                f"Cannot set {self} as global {'admin' if admin else 'moderator'}: "
-                f"{added_by} is not a global admin"
-            )
-            raise BadPermission()
-
-        with db.transaction():
-            query(
-                """
-                UPDATE users
-                SET moderator = TRUE, admin = :admin, visible_mod = :visible
-                WHERE id = :u
-                """,
-                admin=admin,
-                visible=visible,
-                u=self.id,
-            )
-        self.global_admin = admin
-        self.global_moderator = True
-        self.visible_mod = visible
-
-    def remove_moderator(self, *, removed_by: User):
-        """Removes this user's global moderator/admin status, if set."""
-
-        if not removed_by.global_admin:
-            app.logger.warning(
-                f"Cannot remove {self} as global mod/admin: {removed_by} is not an admin"
-            )
-            raise BadPermission()
-
-        with db.transaction():
-            query("UPDATE users SET moderator = FALSE, admin = FALSE WHERE id = :u", u=self.id)
-        self.global_admin = False
-        self.global_moderator = False
-
-    @property
-    def system_user(self):
-        """True iff this is the special SOGS system user created for internal database tasks"""
-        return self.session_id[0:2] == "ff" and self.session_id[2:] == crypto.server_pubkey_hex
-
-
-class SystemUser(User):
-    """
-    User subclasses representing the local system for performing local operations, e.g. from the
-    command line.
-    """
-
-    def __init__(self):
-        super().__init__(session_id="ff" + crypto.server_pubkey_hex)
-
-
 def get_rooms():
     """Get a list of all rooms; does not check permissions."""
     return [Room(row) for row in query("SELECT * FROM rooms ORDER BY token")]
@@ -1494,29 +1119,6 @@ def get_readable_rooms(pubkey=None):
             s=pubkey,
         )
     return [Room(row) for row in result]
-
-
-def get_all_global_moderators():
-    """
-    Returns all global moderators; for internal user only as this doesn't filter out hidden
-    mods/admins.
-
-    Returns a 4-tuple of lists of:
-    - visible mods
-    - visible admins
-    - hidden mods
-    - hidden admins
-    """
-
-    m, hm, a, ha = [], [], [], []
-    for row in query("SELECT * FROM users WHERE moderator"):
-        u = User(row=row)
-        if u.system_user:
-            continue
-        lst = (a if u.global_admin else m) if u.visible_mod else (ha if u.global_admin else hm)
-        lst.append(u)
-
-    return (m, a, hm, ha)
 
 
 def get_deletions_deprecated(room: Room, since):
