+++ sogs/routes.py
@@ -1,10 +1,8 @@
 from flask import abort, request, render_template, Response
 from .web import app
-from . import crypto
-from . import model
-from . import utils
-from . import config
-from . import http
+from . import config, crypto, http, utils
+from .model.room import Room, get_readable_rooms
+from .model.exc import NoSuchRoom
 
 from werkzeug.routing import BaseConverter, ValidationError
 
@@ -20,8 +18,8 @@ class RoomTokenConverter(BaseConverter):
 
     def to_python(self, value):
         try:
-            return model.Room(token=value)
-        except model.NoSuchRoom:
+            return Room(token=value)
+        except NoSuchRoom:
             raise ValidationError()
 
     def to_value(self, value):
@@ -41,7 +39,7 @@ app.url_map.converters['SessionID'] = SessionIDConverter
 
 @app.get("/")
 def serve_index():
-    rooms = model.get_readable_rooms()
+    rooms = get_readable_rooms()
     if len(rooms) == 0:
         return render_template('setup.html')
     return render_template(
